Introduced AbstractRetrievable
